# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RAG Service API
  version: "1.0.0"
  summary: Public API for querying and managing the RAG service.
  description: |
    Versioned public API for external applications integrating with the RAG
    service. All `/api/v1/*` endpoints require an `X-API-Key` header. In
    multi-user deployments, each API key resolves to exactly one user
    workspace; queries, uploads, jobs, conversations, and deletes stay inside
    that workspace boundary.
servers:
  - url: http://127.0.0.1:5000
    description: Local development server
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/models:
    get:
      tags:
        - Models
      operationId: listApiV1Models
      summary: List available models
      description: |
        Returns enabled built-in and custom OpenAI-compatible models that can
        be used in `/api/v1/query`.
      responses:
        "200":
          description: Available models grouped as a flat list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsResponse"
              examples:
                default:
                  value:
                    default_provider: mistral
                    default_model: mistral-medium
                    default_value: mistral:mistral-medium
                    models:
                      - id: mistral-medium
                        name: mistral-medium (Mistral AI)
                        provider: mistral
                        provider_name: Mistral AI
                        value: mistral:mistral-medium
                        is_default: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Model provider configuration is missing or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                modelConfigurationError:
                  value:
                    error: "No models configured in app/default_providers.json."
                    status: model_configuration_error
                    provider_config_file: app/default_providers.json
                    configure_url: /admin/config
...
  /api/v1/agents:
    get:
      tags:
        - Chat Agents
      operationId: listApiV1Agents
      summary: List chat agents visible to the API key
      description: |
        Requires `query` or `agent_manage`. Only agents whose
        `knowledge_base_ids` are a subset of the key's allowlist are returned.
      responses:
        "200":
          description: Visible agents and workspace limits.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatAgentsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - Chat Agents
      operationId: createApiV1Agent
      summary: Create a chat agent
      description: Requires `agent_manage`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatAgentCreateInput"
      responses:
        "201":
          description: Agent created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatAgent"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/KnowledgeBaseNotFound"
        "409":
          description: KB limit, agent count limit, or duplicate name.
  /api/v1/agents/options:
    get:
      tags:
        - Chat Agents
      operationId: getApiV1AgentOptions
      summary: Models, knowledge bases, prompts, and limits for the API key
      description: |
        Requires `query` or `agent_manage`. Useful for building agent
        create/edit forms in external clients.
      responses:
        "200":
          description: Available options.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatAgentOptionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/agents/{agent_id}:
    parameters:
      - $ref: "#/components/parameters/AgentIdPath"
    get:
      tags:
        - Chat Agents
      operationId: getApiV1Agent
      summary: Get one chat agent
      description: |
        Requires `query` or `agent_manage`. An agent referencing a KB outside
        the key's allowlist is hidden with 404.
      responses:
        "200":
          description: Agent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatAgent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ChatAgentNotFound"
    patch:
      tags:
        - Chat Agents
      operationId: updateApiV1Agent
      summary: Update a chat agent
      description: |
        Requires `agent_manage`. All fields are optional; only provided fields
        change. At least one field is required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatAgentUpdateInput"
      responses:
        "200":
          description: Updated agent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatAgent"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ChatAgentNotFound"
        "409":
          description: KB limit or duplicate name.
    delete:
      tags:
        - Chat Agents
      operationId: deleteApiV1Agent
      summary: Delete a chat agent
      description: Requires `agent_manage`.
      responses:
        "200":
          description: Agent deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatAgentDeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ChatAgentNotFound"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: User/workspace API key configured from `/admin/api-keys`. `RAG_API_KEY` remains available as a legacy global compatibility key.
  parameters:
...
    AgentIdPath:
      name: agent_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ChatAgentId"
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing:
              value:
                error: Missing or invalid API key
                status: unauthorized
...
    Forbidden:
      description: API key is valid but lacks the required scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missingScope:
              value:
                error: "API key missing required scope: speech"
                status: forbidden
...
    ValidationError:
      description: Request payload failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidQuery:
              value:
                error: query must have at least 3 characters
                status: validation_error
                field: query
    KnowledgeBaseNotFound:
      description: A selected knowledge base is missing or not authorized to the API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missingOrUnauthorized:
              value:
                error: Knowledge base non trovata
                status: knowledge_base_not_found
...
    ChatAgentNotFound:
      description: Chat agent does not exist or is outside the API key's KB allowlist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missingOrUnauthorized:
              value:
                error: Agent non trovato
                status: chat_agent_not_found
...
  schemas:
    KnowledgeBaseId:
      type: string
      pattern: "^(default|kb_[0-9a-f]{32})$"
      default: default
...
    KnowledgeBase:
      type: object
      required:
        - id
        - name
        - description
        - is_default
        - status
        - created_at
        - updated_at
        - stats
      properties:
        id:
          $ref: "#/components/schemas/KnowledgeBaseId"
        name:
          type: string
        description:
          type: string
        is_default:
          type: boolean
        status:
          type: string
          enum: [active, deleting, delete_failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        delete_error:
          type: string
        stats:
          type: object
          required: [tracked_files, indexed_files, chunks, data_sources]
          properties:
            tracked_files:
              type: integer
              minimum: 0
            indexed_files:
              type: integer
              minimum: 0
            chunks:
              type: integer
              minimum: 0
            data_sources:
              type: integer
              minimum: 0
    ChatAgentId:
      type: string
      pattern: "^agent_[0-9a-f]{32}$"
    ChatAgentPromptRef:
      type: object
      required: [id, scope]
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 200
        scope:
          type: string
          enum: [personal, shared]
      additionalProperties: false
    ChatAgentAvailabilityIssue:
      type: object
      required: [code, field, message]
      properties:
        code:
          type: string
          examples:
            - model_unavailable
            - knowledge_base_missing
            - knowledge_base_inactive
            - knowledge_base_limit_exceeded
            - prompt_missing
            - prompt_inactive
        field:
          type: string
        message:
          type: string
    ChatAgent:
      type: object
      required:
        - id
        - name
        - description
        - provider_id
        - model_id
        - knowledge_base_ids
        - prompt_ref
        - created_at
        - updated_at
        - available
        - issues
      properties:
        id:
          $ref: "#/components/schemas/ChatAgentId"
        name:
          type: string
        description:
          type: string
        provider_id:
          type: string
        model_id:
          type: string
        knowledge_base_ids:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/KnowledgeBaseId"
        prompt_ref:
          $ref: "#/components/schemas/ChatAgentPromptRef"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        available:
          type: boolean
          description: True when the agent can currently run (no availability issues).
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ChatAgentAvailabilityIssue"
    ChatAgentsResponse:
      type: object
      required: [agents, limits, capabilities]
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/ChatAgent"
        limits:
          type: object
          required: [max_chat_agents, max_query_knowledge_bases]
          properties:
            max_chat_agents:
              type: integer
              minimum: 0
            max_query_knowledge_bases:
              type: integer
              minimum: 1
        capabilities:
          type: object
          required: [can_manage]
          properties:
            can_manage:
              type: boolean
    ChatAgentCreateInput:
      type: object
      required:
        - name
        - provider_id
        - model_id
        - knowledge_base_ids
        - prompt_ref
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
        provider_id:
          type: string
          minLength: 1
          maxLength: 120
        model_id:
          type: string
          minLength: 1
          maxLength: 200
        knowledge_base_ids:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/KnowledgeBaseId"
        prompt_ref:
          $ref: "#/components/schemas/ChatAgentPromptRef"
      additionalProperties: false
    ChatAgentUpdateInput:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
        provider_id:
          type: string
          minLength: 1
          maxLength: 120
        model_id:
          type: string
          minLength: 1
          maxLength: 200
        knowledge_base_ids:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/KnowledgeBaseId"
        prompt_ref:
          $ref: "#/components/schemas/ChatAgentPromptRef"
      additionalProperties: false
    ChatAgentDeleteResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ChatAgentOptionsResponse:
      type: object
      required:
        - models
        - default_provider
        - default_model
        - knowledge_bases
        - prompts
        - limits
        - capabilities
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelInfo"
        default_provider:
          type: string
        default_model:
          type: string
        knowledge_bases:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeBase"
        prompts:
          type: array
          items:
            type: object
            required: [id, name, scope, is_active]
            properties:
              id:
                type: string
              name:
                type: string
              scope:
                type: string
                enum: [personal, shared]
              is_active:
                type: boolean
        limits:
          type: object
          required: [max_chat_agents, max_query_knowledge_bases]
          properties:
            max_chat_agents:
              type: integer
              minimum: 0
            max_query_knowledge_bases:
              type: integer
              minimum: 1
        capabilities:
          type: object
          required: [can_manage]
          properties:
            can_manage:
              type: boolean
    ErrorResponse:
      type: object
      required:
        - error
        - status
      properties:
        error:
          type: string
        status:
          type: string
          examples:
            - validation_error
            - unauthorized
            - rate_limited
            - server_error
            - model_configuration_error
        field:
          type: string
          description: Field that failed validation, when available.
        retry_after:
          type: integer
          minimum: 1
          description: Seconds to wait before retrying after rate limiting.
        provider_config_file:
          type: string
          description: Provider model configuration file path.
        configure_url:
          type: string
          description: Admin page used to fix runtime or model configuration.
      additionalProperties: true
...
    ModelsResponse:
      type: object
      required:
        - models
      properties:
        default_provider:
          type: string
          description: Provider selected by runtime configuration when a query omits provider.
          examples:
            - mistral
        default_model:
          type: string
          description: Model selected by runtime configuration when a query omits model.
          examples:
            - mistral-medium
        default_value:
          type: string
          description: Compact `provider:model` value for the runtime default.
          examples:
            - mistral:mistral-medium
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelInfo"
    ModelInfo:
      type: object
      required:
        - id
        - name
        - provider
        - provider_name
        - value
      properties:
        id:
          type: string
          examples:
            - mistral-medium
        name:
          type: string
          examples:
            - mistral-medium (Mistral AI)
        provider:
          type: string
          examples:
            - mistral
        provider_name:
          type: string
          examples:
            - Mistral AI
        value:
          type: string
          description: Compact UI value in `provider:model` format.
          examples:
            - mistral:mistral-medium
        is_default:
          type: boolean
          description: True when this model is the configured runtime default.
          examples:
            - true
...